sLLM/requirements.txt
black
pre-commit
fastapi
uvicorn[standard]
python-dotenv
langchain_openai